pull request from AndrijanaStefanovic

+++ src/main/java/com/example/repository/TableRepository.java
@@ -3,13 +3,14 @@ package com.example.repository;
 import java.util.List;
 
 import org.springframework.data.repository.Repository;
+
 import com.example.model.TableRestaurant;
 
-public interface TableRepository extends Repository<TableRestaurant, Long> {
+public interface TableRepository extends Repository<TableRestaurant, Integer> {
 	
 	public TableRestaurant save(TableRestaurant entity);
 	
-	public TableRestaurant findOne(long id);
+	public TableRestaurant findOne(int id);
 	
 	public List<TableRestaurant> findByRestaurant(long restaurant);
 }
